fix: Bitget API互換性修正とバックテスト制限対応
修正内容:
- config.py: entry_range/stop_range を int → float に変更
- bitget_exchange.py: timeframe '2h' → '1h' (Bitgetは2hをサポートしない)
- bitget_exchange.py: エラーログに詳細パラメータ追加
- price_data_management.py: ログメッセージを動的に取引所名を表示

重要な発見:
**Bitgetの制限**: 過去データ取得が約8日間のみ
- バックテスト（2024年からのデータ）には対応不可
- キャッシュデータ（Bybitで取得済み）を活用する必要あり
- 本番取引のみBitgetで実行可能

今後の対応:
1. バックテストはBybitキャッシュデータを使用
2. ホットテスト・本番取引のみBitgetを使用
3. またはBybitとBitgetを併用する構成を検討

+++ src/config.py
@@ -105,9 +105,9 @@ class Config:
         エントリー範囲を取得します.
 
         Returns:
-            int: エントリー範囲
+            float: エントリー範囲
         """
-        return int(cls.config['RiskManagement']['entry_range'])
+        return float(cls.config['RiskManagement']['entry_range'])
 
     @classmethod
     def get_stop_range(cls):
@@ -115,9 +115,9 @@ class Config:
         ストップ範囲を取得します.
 
         Returns:
-            int: ストップ範囲
+            float: ストップ範囲
         """
-        return int(cls.config['RiskManagement']['stop_range'])
+        return float(cls.config['RiskManagement']['stop_range'])
 
     @classmethod
     def get_stop_AF(cls):
